✨ feat(account): add account options and label retrieval
