--- src/lib/accountOptions.ts
@@ -1,5 +1,0 @@
-export const accountOptions = [
-  { value: 'cash', label: '現金' },
-  { value: 'bank', label: '銀行' },
-  { value: 'credit', label: '信用卡' },
-];
